Make blocksize thread_local
This change uses the recently enabled C11 and declares
blocksize thread local. The ext2 blocksize variable is
the only global variable in genext2fs.c and exposing it
for change in abuild would otherwise break multiple threads
with different blocksizes.

With thread_local, each thread using this variable obtains
a fresh copy initialized with the default size of 1024.

Signed-off-by: Markus Theil <markus.theil@secunet.com>